+++ packages/components/src/shared/bricks/vbox.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from "react";
 import { useBrickStyle } from "../hooks/use-brick-style";
-import { type Manifest, manifest } from "@upstart.gg/sdk/bricks/manifests/container.manifest";
+import { type Manifest, manifest } from "@upstart.gg/sdk/shared/bricks/manifests/vbox.manifest";
 import EditableBrickWrapper from "~/editor/components/EditableBrick";
 import type { BrickProps } from "@upstart.gg/sdk/shared/bricks/props/types";
 import { useDatasource } from "../hooks/use-datasource";
@@ -9,10 +9,11 @@ import BrickWrapper from "../components/BrickWrapper";
 import { tx } from "@upstart.gg/style-system/twind";
 import { Droppable } from "@hello-pangea/dnd";
 
-const Container = forwardRef<HTMLDivElement, BrickProps<Manifest>>(({ brick, editable }, ref) => {
+const Vbox = forwardRef<HTMLDivElement, BrickProps<Manifest>>(({ brick, editable }, ref) => {
   const props = brick.props;
 
   const styles = useBrickStyle<Manifest>(brick);
+  const classes = Object.values(styles);
   const ds = useDatasource(props.datasource, manifest.datasource);
 
   // If this container is Dynamic
@@ -36,15 +37,16 @@ const Container = forwardRef<HTMLDivElement, BrickProps<Manifest>>(({ brick, edi
 
   if (editable) {
     return (
-      <div className={tx("flex-1 flex overflow-hidden relative", Object.values(styles))} ref={ref}>
+      <div className={tx("flex flex-1 flex-col relative", ...classes)} ref={ref}>
         <Droppable droppableId={brick.id} type="brick" direction="vertical">
           {(droppableProvided, droppableSnapshot) => (
             <div
               {...droppableProvided.droppableProps}
               ref={droppableProvided.innerRef}
               className={tx(
-                "flex-1",
+                "flex-1 flex flex-col relative",
                 droppableSnapshot.isDraggingOver && "!outline !outline-2 !outline-orange-300",
+                ...classes,
               )}
             >
               {props.$children?.length > 0 ? (
@@ -86,4 +88,4 @@ const Container = forwardRef<HTMLDivElement, BrickProps<Manifest>>(({ brick, edi
   );
 });
 
-export default Container;
+export default Vbox;