+++ packages/sdk/src/shared/bricks/props/prop.ts
@@ -5,6 +5,5 @@ import type { Prop } from "./types";
  * Helper function for creating props with additional metadata
  */
 export function prop<T extends TSchema>({ schema, ...rest }: Prop<T>): T {
-  Object.assign(schema, rest);
-  return schema;
+  return Object.assign({}, schema, rest);
 }