+++ packages/components/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @upstart.gg/components
 
+## 0.0.115
+
+### Patch Changes
+
+- [#355](https://github.com/upstart-gg/upstart/pull/355) [`cf9f387`](https://github.com/upstart-gg/upstart/commit/cf9f3870d3040093015632e252a9f5ebf574f4d7) Thanks [@mattallty](https://github.com/mattallty)! - Improve bricks layout
+
+- Updated dependencies [[`cf9f387`](https://github.com/upstart-gg/upstart/commit/cf9f3870d3040093015632e252a9f5ebf574f4d7)]:
+  - @upstart.gg/sdk@0.0.115
+  - @upstart.gg/style-system@0.0.115
+
+## 0.0.114
+
+### Patch Changes
+
+- [#353](https://github.com/upstart-gg/upstart/pull/353) [`e709fcb`](https://github.com/upstart-gg/upstart/commit/e709fcbe493df3808ea1b1996d5ce6484f2cde8d) Thanks [@mattallty](https://github.com/mattallty)! - More elements
+
+- Updated dependencies [[`e709fcb`](https://github.com/upstart-gg/upstart/commit/e709fcbe493df3808ea1b1996d5ce6484f2cde8d)]:
+  - @upstart.gg/style-system@0.0.114
+  - @upstart.gg/sdk@0.0.114
+
 ## 0.0.113
 
 ### Patch Changes